v2/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyJWT>=2.8.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
